+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # 0.0.2 ()
 
-* [ADDED] Function to list defcomponent ignored keys []()
+* [ADDED] Function to list defwebcomponent ignored keys []()
 * [ADDED] Functions to check for Web Components technologies support []()
 * [ADDED] Enhanced `host` syntax to support attributes []()
 * [ADDED] Extended `style` syntax to support media queries []()
 * [UPDATED] Renamed `base-type` to `host`
+* [UPDATED] Renamed handlers (created-fn -> on-created, entered-view-fn -> on-attached, left-view-fn -> on-detached)
 * [UPDATED] Moved examples to Custom Component with real live view support
 * And a bunch of [bugs fixed](https://github.com/jeluard/lucuma/issues?labels=bug&milestone=1&page=1&state=closed).